+++ ukat/data/fetch.py
@@ -195,6 +195,26 @@ fetch_mtr_philips = _make_fetcher('fetch_mtr_philips',
                                   ['252fcc0d67feb6ea3a55b850eb1f4477'],
                                   doc='Downloading Philips MT data')
 
+fetch_pc_left_philips = _make_fetcher('fetch_pc_left_philips',
+                                      pjoin(ukat_home, 'pc_left_philips'),
+                                      'https://zenodo.org/record/5655752'
+                                      '/files/',
+                                      ['philips_pc_left.zip'],
+                                      ['philips_pc_left.zip'],
+                                      ['97550f62e0a6c9cc0bc4ac2f1c52a7ea'],
+                                      unzip=True,
+                                      doc='Downloading Philips PC Left data')
+
+fetch_pc_right_philips = _make_fetcher('fetch_pc_right_philips',
+                                       pjoin(ukat_home, 'pc_right_philips'),
+                                       'https://zenodo.org/record/5655752'
+                                       '/files/',
+                                       ['philips_pc_right.zip'],
+                                       ['philips_pc_right.zip'],
+                                       ['d5bcc1d70ff43ecec4f77889099d7055'],
+                                       unzip=True,
+                                       doc='Downloading Philips PC Right data')
+
 fetch_t1_philips_1 = _make_fetcher('fetch_t1_philips_1',
                                    pjoin(ukat_home, 't1_philips_1'),
                                    'https://zenodo.org/record/4762285/files/',
@@ -220,6 +240,23 @@ fetch_t1_philips_2 = _make_fetcher('fetch_t1_philips_2',
                                     '8b16dd5f00abde1aff6845ecd38f74f7',
                                     'b1bc6c2f6c43e26f4a1d27868eb93df3'],
                                    doc='Downloading Philips T1 dataset 2')
+
+fetch_t1_molli_philips = _make_fetcher('fetch_t1_molli_philips',
+                                       pjoin(ukat_home, 't1_molli_philips'),
+                                       'https://zenodo.org/record/5846750/'
+                                       'files/',
+                                       ['01101_WIP_Cor_T1_MOLLI_e1.json',
+                                        '01101_WIP_Cor_T1_MOLLI_e1.nii.gz',
+                                        'ti.csv'],
+                                       ['01101_WIP_Cor_T1_MOLLI_e1.json',
+                                        '01101_WIP_Cor_T1_MOLLI_e1.nii.gz',
+                                        'ti.csv'],
+                                       ['31543856a34b1696b5b56af4fb2427c8',
+                                        '42c1dc5c49326fdaeae97d0473b83984',
+                                        '82bf71c1dab4a490f7cdc66d887afb94'],
+                                       doc='Downloading Philips T1 MOLLI '
+                                           'dataset'
+                                       )
 
 fetch_t1w_philips = _make_fetcher('fetch_t1w_philips',
                                   pjoin(ukat_home, 't1w_philips'),
@@ -276,15 +313,6 @@ fetch_t2w_philips = _make_fetcher('fetch_t2w_philips',
                                    '.nii.gz'],
                                   ['276b904142677026a04659505d923134'],
                                   doc='Downloading Philips T2W data')
-
-fetch_mtr_philips = _make_fetcher('fetch_mtr_philips',
-                                  pjoin(ukat_home, 'mtr_philips'),
-                                  'https://zenodo.org/record/5101394/'
-                                  'files/',
-                                  ['Cor_2D_MTR_BH_3201.nii.gz'],
-                                  ['Cor_2D_MTR_BH_3201.nii.gz'],
-                                  ['252fcc0d67feb6ea3a55b850eb1f4477'],
-                                  doc='Downloading Philips MT data')
 
 fetch_tsnr_high_philips = _make_fetcher('fetch_tsnr_high_philips',
                                         pjoin(ukat_home, 'tsnr_high_philips'),
@@ -357,6 +385,16 @@ def get_fnames(name):
         fnames = sorted(glob.glob(pjoin(folder, '*')))
         return fnames
 
+    elif name == 'phase_contrast_left_philips':
+        files, folder = fetch_pc_left_philips()
+        fnames = sorted(glob.glob(pjoin(folder, '*')))
+        return fnames
+
+    elif name == 'phase_contrast_right_philips':
+        files, folder = fetch_pc_right_philips()
+        fnames = sorted(glob.glob(pjoin(folder, '*')))
+        return fnames
+
     elif name == 't1_philips_1':
         files, folder = fetch_t1_philips_1()
         fnames = sorted(glob.glob(pjoin(folder, '*')))
@@ -364,6 +402,11 @@ def get_fnames(name):
 
     elif name == 't1_philips_2':
         files, folder = fetch_t1_philips_2()
+        fnames = sorted(glob.glob(pjoin(folder, '*')))
+        return fnames
+
+    elif name == 't1_molli_philips':
+        files, folder = fetch_t1_molli_philips()
         fnames = sorted(glob.glob(pjoin(folder, '*')))
         return fnames
 
@@ -396,9 +439,6 @@ def get_fnames(name):
         files, folder = fetch_t2w_philips()
         fnames = sorted(glob.glob(pjoin(folder, '*')))
         return fnames
-
-    elif name == 'mtr_philips':
-        files, folder = fetch_mtr_philips()
 
     elif name == 'tsnr_high_philips':
         files, folder = fetch_tsnr_high_philips()
@@ -601,9 +641,88 @@ def mtr_philips():
     fnames = get_fnames('mtr_philips')
     nii_path = [f for f in fnames if f.endswith('.nii.gz')][0]
     raw = nib.load(nii_path)
-    data = raw.get_fdata()
+    data = np.squeeze(raw.get_fdata())
     affine = raw.affine
     return data, affine
+
+
+def phase_contrast_left_philips():
+    """Fetches pc_left/philips dataset
+        Returns
+        -------
+        numpy.ndarray
+            image data
+        numpy.ndarray
+            boolean mask for image data
+        numpy.ndarray
+            affine matrix for image data
+        float
+            velocity encoding of the phase contrast scan data
+    """
+    fnames = get_fnames('phase_contrast_left_philips')
+    magnitude = []
+    phase = []
+    velocity_encoding = 100
+    for file in fnames:
+        if ((file.endswith(".nii.gz") and "_ph_" in file) or
+           file.endswith("_ph.nii.gz")):
+            # Load NIfTI and only save the phase data
+            data = nib.load(file)
+            phase.append(np.squeeze(data.get_fdata()))
+
+        elif file.endswith(".nii.gz") and "mask_" in file:
+            mask = np.squeeze(nib.load(file).get_fdata())
+
+        elif file.endswith(".nii.gz"):
+            # Load NIfTI and only save the magnitude data
+            data = nib.load(file)
+            magnitude.append(np.squeeze(data.get_fdata()[..., 0]))
+
+    # Move cardiac cycle dimension to 3rd (and last) dimension
+    magnitude = np.moveaxis(np.array(magnitude), 0, -1)
+    phase = np.moveaxis(np.array(phase), 0, -1)
+
+    return magnitude, phase, mask, data.affine, velocity_encoding
+
+
+def phase_contrast_right_philips():
+    """Fetches pc_right/philips dataset
+        Returns
+        -------
+        numpy.ndarray
+            image data
+        numpy.ndarray
+            boolean mask for image data
+        numpy.ndarray
+            affine matrix for image data
+        float
+            velocity encoding of the phase contrast scan data
+    """
+    fnames = get_fnames('phase_contrast_right_philips')
+    magnitude = []
+    phase = []
+    velocity_encoding = 100
+    for file in fnames:
+
+        if ((file.endswith(".nii.gz") and "_ph_" in file) or
+           file.endswith("_ph.nii.gz")):
+            # Load NIfTI and only save the phase data
+            data = nib.load(file)
+            phase.append(np.squeeze(data.get_fdata()))
+
+        elif file.endswith(".nii.gz") and "mask_" in file:
+            mask = np.squeeze(nib.load(file).get_fdata())
+
+        elif file.endswith(".nii.gz"):
+            # Load NIfTI and only save the magnitude data
+            data = nib.load(file)
+            magnitude.append(np.squeeze(data.get_fdata()[..., 0]))
+
+    # Move cardiac cycle dimension to 3rd (and last) dimension
+    magnitude = np.moveaxis(np.array(magnitude), 0, -1)
+    phase = np.moveaxis(np.array(phase), 0, -1)
+
+    return magnitude, phase, mask, data.affine, velocity_encoding
 
 
 def t1_philips(dataset_id):
@@ -676,6 +795,26 @@ def t1_philips(dataset_id):
         affine = magnitude_img.affine
 
         return magnitude, phase, affine, inversion_list, tss
+
+
+def t1_molli_philips():
+    """Fetches Philips MOLLI T1 dataset
+    Returns
+    -------
+    numpy.ndarray
+        image data
+    numpy.ndarray
+        affine matrix for image data
+    numpy.ndarray
+        inversion times in seconds
+    """
+    fnames = get_fnames('t1_molli_philips')
+
+    data = nib.load(fnames[1])
+    image = data.get_fdata()
+    inversion_list = np.loadtxt(fnames[2])
+
+    return image, data.affine, inversion_list / 1000
 
 
 def t1w_volume_philips():